requirements.txt
django-filter

faker
Django==2.1.3
Pillow==4.0.0
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8

psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
